fixed some architecture errors fixing delete buildings button fixed small errors replaces arraylists by vectors

+++ KohtopaDesktop/KohtopaDesktop/src/gui/actions/BuildingRemoveAction.java
@@ -1,10 +1,15 @@
 package gui.actions;
 
+import Language.Language;
+import data.entities.Building;
+import gui.JActionButton;
 import gui.Main;
 import gui.addremovetab.IBuildingListContainer;
 import java.awt.event.ActionEvent;
+import java.sql.SQLException;
+import java.util.Vector;
 import javax.swing.Icon;
-import javax.swing.JComponent;
+import javax.swing.JOptionPane;
 
 /**
  *
@@ -18,7 +23,18 @@ public class BuildingRemoveAction extends AbstractIconAction {
 
 	@Override
 	public void actionPerformed(ActionEvent e) {
-		Object[] selected = ((IBuildingListContainer)((JComponent)e.getSource()).getRootPane()).getSelectedBuildings();
-		Main.getDataObject().deleteSelectedBuildings(selected);
+		Vector<Integer> selected = new Vector<Integer>();
+		for (Object building : ((IBuildingListContainer) ((JActionButton) e.getSource()).getRoot()).getSelectedBuildings()) {
+			selected.add(((Building) building).getId());
+		}
+		if (JOptionPane.showConfirmDialog(
+				null,
+				Language.getString("confirmDelete") + " " + selected.size() + " " + Language.getString("buildings"), //TODO p000 change building(s) to building if 1 else buildings
+				Language.getString("confirm"),
+				JOptionPane.YES_NO_CANCEL_OPTION,
+				JOptionPane.QUESTION_MESSAGE) == JOptionPane.YES_OPTION) {
+			//delete images
+			Main.getDataObject().deleteSelectedBuildings(selected);
+		}
 	}
 }
